# Lazy load some big assets and optimize the bundle size