--- a/package.json
+++ b/package.json
@@ -82,7 +82,8 @@
     "vue": "^2.6.10",
     "vue-electron": "^1.0.6",
     "vue-router": "^3.0.7",
-    "vuex": "^3.1.1"
+    "vuex": "^3.1.1",
+    "webfontloader": "^1.6.28"
   },
   "devDependencies": {
     "@babel/core": "^7.6.4",
@@ -130,6 +131,7 @@
     "file-loader": "^4.1.0",
     "git-revision-webpack-plugin": "^3.0.4",
     "html-webpack-plugin": "^3.2.0",
+    "imports-loader": "^0.8.0",
     "karma": "^4.1.0",
     "karma-chai": "^0.1.0",
     "karma-coverage": "^2.0.1",