--- a/src/muya/lib/parser/render/sequence.js
+++ b/src/muya/lib/parser/render/sequence.js
@@ -1,5 +1,5 @@
 
-import Diagram from '../../assets/libs/sequence-diagram'
+import Diagram from '../../assets/libs/sequence-diagram-snap'
 import '../../assets/styles/sequence-diagram.css'
 
 export default Diagram