| Q                | A
| ---------------- | ---
| Bug fix?         | yes
| Fixed tickets    | #1033
| License          | MIT

### Description

- [x] Lazy load sequence
- [x] Lazy load vega
- [x] Lazy load mermaid
- [x] Lazy load flowchart
- [x] Update vega-embed etc ~~There still some error when render vega-lite after updated.~~
- [x] Replace raphael with snap.svg(TOTAL render.js size 9.18M to 8.46M)

[Description of the bug or feature]


Render.js | First Meaningful Paint(3 times)
-- | --
Full load | 2768.3 ms、2738.8 ms、2751.2 ms
Without mermaid module | 2180.7ms、 2314.9 ms、 2190.9 ms
Without Vega-embed | 2368.3 ms、2782.3 ms、 2608.8 ms
Without sequence | 2212.1 ms、 2214.1 ms、 2464.4 ms
Without flowchart | 2328.6 ms、 2703.8 ms、2846.4 ms
Lazy load all diagrams | 2042.0 ms、 1855.6 ms、2088.5 ms



--

#### Please, don't submit `/dist` files with your PR!

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/marktext/marktext/1481)
<!-- Reviewable:end -->